ssh.socket: adjust unit for socket activation by default
Add RequiredBy=ssh.service and Before=ssh.service to ensure correct
ordering with ssh.service. This was previously done with a drop-in
override for ssh.service, but we can specify these in ssh.socket
directly.

Also set FreeBind=yes so that if we generate ListenStream= entries with
specific IP addresses, the unit is more robust and will not fail if it
is started before the addresses are actually configured.