Add a systemd generator for socket activation
This patch adds a binary, sshd-socket-generator, which reads
/etc/ssh/sshd_config (and /etc/ssh/sshd_config.d/*), and generates a
drop-in ssh.socket.d/addresses.conf with ListenStream= created according
to Port and ListenAddress entries.

This means that users of systemd socket activated sshd can still
modify their sshd config as usual, and have it take effect in
ssh.socket.

The generator is run on systemctl daemon-reload, so the workflow
becomes:

 * Modify /etc/ssh/sshd_config.d/foo.conf
 * systemctl daemon-reload
 * systemctl restart ssh.socket

Author: Nick Rosbrook <enr0n@ubuntu.com>
Last-Update: 2024-07-30

Patch-Name: sshd-socket-generator.patch